+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'kong'
+rootProject.name = 'externalcreds'
