+++ src/main/java/bio/terra/externalcreds/ExternalCredsApplication.java
@@ -1,12 +1,12 @@
-package bio.terra.kong;
+package bio.terra.externalcreds;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication(scanBasePackages = {"bio.terra.common.logging"})
-public class KongApplication {
+public class ExternalCredsApplication {
 
   public static void main(String[] args) {
-    SpringApplication.run(KongApplication.class, args);
+    SpringApplication.run(ExternalCredsApplication.class, args);
   }
 }
