+++ src/test/java/bio/terra/externalcreds/KongApplicationTests.java
@@ -1,10 +1,10 @@
-package bio.terra.kong;
+package bio.terra.externalcreds;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class KongApplicationTests {
+class ExternalCredsApplicationTests {
 
   @Test
   void contextLoads() {}
